fixed kivy reloder on windows issues were: original issue: when you save and force a reload on windows, keyboardinterrupt was not being caught by new spawned child processes
this is because windows Process is not the same as os.spawnv on linux. (children do not listen to parent's KeyboardInterrupt when parent has exited before child)

the fix is to keep the host open but doing nothing and only "listening" for keyboardinterrupt so it will be send to a future child process.

the fix is to set some argv that is the host Python  pid for 2 things:

#1: intermediate children will exit if they know their argv has the hostPID and they will be replaced with another reloaded child process

#2: the final child that is closed with the "X" button on windows has an on_request_close bound to it that tasks windows to taskkill the original host python. Kinda chicken and egg? but it works on my machine....

#3: the final child will also respond to KeyboardInterrupt because parent is still active.

other issues were:

- FIXED: new windows being created and not destroyed when reloading

- FIXED: when pressing X on the final child window, the original host python would not exit and would be guaranteed leaked memory.